+++ CHANGES.rst
@@ -7,3 +7,6 @@ Changelog
 
 - Initial release.
   [mpeeters, sgeulette]
+- Replaced collective.z3cform.chosen widget by collective.z3cform.select2.
+  Must remove "chosen" packages in next release.
+  [sgeulette]